third_party/blink/web_tests/images/animated-images-use-counter-004.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<script>
const kAnimatedImageUsedMoreThanOnce = 5605;

async_test(function(t) {
  internals.clearUseCounter(document, kAnimatedImageUsedMoreThanOnce);

  let png1 = document.createElement("img");
  png1.src = "resources/png-simple.png";
  document.body.appendChild(png1);
  let png2 = document.createElement("img");
  png2.src = "resources/png-simple.png";
  document.body.appendChild(png2);

  onload = t.step_func(() => {
    requestAnimationFrame(t.step_func(() => {
      requestAnimationFrame(t.step_func_done(() => {
        assert_false(internals.isUseCounted(document, kAnimatedImageUsedMoreThanOnce));
      }));
    }));
  });
}, 'Check that kAnimatedImageUsedMoreThanOnce is NOT counted with duplicated images not animated');
</script>
</body>
